# packages/syndb-api-client/syndb-api-client-0.1.0.tar.gz/syndb-api-client-0.1.0/syndb_api/neurodata/metadata/schema.py
from typing import Optional
from uuid import UUID

from pydantic import Field

from syndb_api.common.schema import SchemaModel


class CommonMetadata(SchemaModel):
    authorized_group_names: Optional[list[str]] = Field(default_factory=list)


class IdResponse(SchemaModel):
    id: UUID


class Id2LabelResponse(SchemaModel):
    uuid_to_label: dict[UUID, str]